azure-functions/requirements.txt
azure-functions>=1.18.0
azure-storage-blob>=12.19.0
Pillow>=10.2.0
stripe>=8.0.0
sendgrid>=6.11.0
requests>=2.31.0
python-dotenv>=1.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0